test(l1): move the new trie tests into the `test` crate
This branch added its unit tests inline, as `#[cfg(test)]` modules beside
the code. The repository keeps tests in the `ethrex-test` crate instead,
one file per area under `test/tests/<subsystem>/`, aggregated through a
`mod.rs`, so put them where they belong:

  crates/common/trie/path_cursor.rs  -> test/tests/trie/path_cursor_tests.rs
  crates/storage/layering.rs         -> test/tests/storage/layering_tests.rs

Test counts move accordingly and nothing is lost: `ethrex-trie` goes from
70 back to its pre-branch 61, `ethrex-storage` from 92 to 91, and the
integration binary picks up the same ten (trie 67 to 76, storage 24 to
25). The bodies are unchanged apart from unwrapping the `mod tests`
indentation and importing through the public crate root.

`apply_prefix_bytes` is now re-exported next to `apply_prefix`, which it
is the borrowing counterpart of, so its test can reach it from outside the
crate. The pre-existing inline tests in `layering.rs` stay put: they cover
private items and predate this branch.
