+++ test/tests/storage/mod.rs
@@ -1,5 +1,6 @@
 mod deferred_persistence_tests;
 mod fcu_race_tests;
+mod layering_tests;
 mod storage_batch_tests;
 mod store_tests;
 mod trie_db_tests;
