+++ test/tests/trie/mod.rs
@@ -1,4 +1,5 @@
 mod nibbles_tests;
+mod path_cursor_tests;
 mod trie_iter_tests;
 mod trie_tests;
 mod verify_range_tests;
